(PC-41905)[PRO] test: rework tests after review
